+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.1] - 2021-11-22
 
 - Removed 'hostname' argument for `load_ssh_config` method in `SSHConfigData`.
+- Added `is_active` method to `SSH`.
 
 ## [0.1.0] - 2021-09-22